Переместил файлы тестов
