+++ public-transport-timetable-biz/src/commonTest/kotlin/space/rybakov/timetable/biz/validation/BizValidationCreateTest.kt
@@ -1,4 +1,5 @@
-package space.rybakov.timetable.biz.validation.validation
+package space.rybakov.timetable.biz.validation
+
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import space.rybakov.timetable.backend.repository.inmemory.TripRepoStub
@@ -7,10 +8,11 @@ import space.rybakov.timetable.common.TimetableCorSettings
 import space.rybakov.timetable.common.models.TimetableCommand
 import kotlin.test.Test
 
+// TODO-validation-5: смотрим пример теста валидации, собранного из тестовых функций-оберток
 @OptIn(ExperimentalCoroutinesApi::class)
-class BizValidationUpdateTest {
+class BizValidationCreateTest {
 
-    private val command = TimetableCommand.UPDATE
+    private val command = TimetableCommand.CREATE
     private val settings by lazy {
         TimetableCorSettings(
             repoTest = TripRepoStub()
@@ -18,21 +20,15 @@ class BizValidationUpdateTest {
     }
     private val processor by lazy { TimetableTripProcessor(settings) }
 
-    @Test fun correctTitle() = validationNameCorrect(command, processor)
-    @Test fun trimTitle() = validationNameTrim(command, processor)
-    @Test fun emptyTitle() = validationNameEmpty(command, processor)
-    @Test fun badSymbolsTitle() = validationNameSymbols(command, processor)
+    @Test fun correctName() = validationNameCorrect(command, processor)
+    @Test fun trimName() = validationNameTrim(command, processor)
+    @Test fun emptyName() = validationNameEmpty(command, processor)
+    @Test fun badSymbolsName() = validationNameSymbols(command, processor)
 
     @Test fun correctDescription() = validationDescriptionCorrect(command, processor)
     @Test fun trimDescription() = validationDescriptionTrim(command, processor)
     @Test fun emptyDescription() = validationDescriptionEmpty(command, processor)
     @Test fun badSymbolsDescription() = validationDescriptionSymbols(command, processor)
-
-    @Test fun correctId() = validationIdCorrect(command, processor)
-    @Test fun trimId() = validationIdTrim(command, processor)
-    @Test fun emptyId() = validationIdEmpty(command, processor)
-    @Test fun badFormatId() = validationIdFormat(command, processor)
-
 
 }
