+++ public-transport-timetable-biz/src/commonTest/kotlin/space/rybakov/timetable/biz/validation/BizValidationSearchTest.kt
@@ -1,4 +1,4 @@
-package space.rybakov.timetable.biz.validation.validation
+package space.rybakov.timetable.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
