+++ public-transport-timetable-biz/src/commonTest/kotlin/space/rybakov/timetable/biz/validation/ValidationBadDescription.kt
@@ -1,4 +1,4 @@
-package space.rybakov.timetable.biz.validation.validation
+package space.rybakov.timetable.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
@@ -9,11 +9,9 @@ import space.rybakov.timetable.stubs.TimetableTripStub
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
-
 private val stub = TimetableTripStub.get()
-
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationNameCorrect(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
+fun validationDescriptionCorrect(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
     val ctx = TimetableContext(
         command = command,
         state = TimetableState.NONE,
@@ -28,38 +26,38 @@ fun validationNameCorrect(command: TimetableCommand, processor: TimetableTripPro
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(TimetableState.FAILING, ctx.state)
-    assertEquals("abc", ctx.tripValidated.name)
+    assertEquals("abc", ctx.tripValidated.description)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationNameTrim(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
+fun validationDescriptionTrim(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
     val ctx = TimetableContext(
         command = command,
         state = TimetableState.NONE,
         workMode = TimetableWorkMode.TEST,
         tripRequest = TimetableTrip(
             id = stub.id,
-            name = " \n\t abc \t\n ",
-            description = "abc",
+            name = "abc",
+            description = " \n\tabc \n\t",
             tripType = TimetableDirection.FORWARD,
         ),
     )
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(TimetableState.FAILING, ctx.state)
-    assertEquals("abc", ctx.tripValidated.name)
+    assertEquals("abc", ctx.tripValidated.description)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationNameEmpty(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
+fun validationDescriptionEmpty(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
     val ctx = TimetableContext(
         command = command,
         state = TimetableState.NONE,
         workMode = TimetableWorkMode.TEST,
         tripRequest = TimetableTrip(
             id = stub.id,
-            name = "",
-            description = "abc",
+            name = "abc",
+            description = "",
             tripType = TimetableDirection.FORWARD,
         ),
     )
@@ -67,20 +65,20 @@ fun validationNameEmpty(command: TimetableCommand, processor: TimetableTripProce
     assertEquals(1, ctx.errors.size)
     assertEquals(TimetableState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("name", error?.field)
-    assertContains(error?.message ?: "", "name")
+    assertEquals("description", error?.field)
+    assertContains(error?.message ?: "", "description")
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationNameSymbols(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
+fun validationDescriptionSymbols(command: TimetableCommand, processor: TimetableTripProcessor) = runTest {
     val ctx = TimetableContext(
         command = command,
         state = TimetableState.NONE,
         workMode = TimetableWorkMode.TEST,
         tripRequest = TimetableTrip(
-            id = TimetableTripId("123"),
-            name = "!@#$%^&*(),.{}",
-            description = "abc",
+            id = stub.id,
+            name = "abc",
+            description = "!@#$%^&*(),.{}",
             tripType = TimetableDirection.FORWARD,
         ),
     )
@@ -88,6 +86,6 @@ fun validationNameSymbols(command: TimetableCommand, processor: TimetableTripPro
     assertEquals(1, ctx.errors.size)
     assertEquals(TimetableState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("name", error?.field)
-    assertContains(error?.message ?: "", "name")
+    assertEquals("description", error?.field)
+    assertContains(error?.message ?: "", "description")
 }
