--- public-transport-timetable-biz/src/commonTest/kotlin/space/rybakov/timetable/biz/validation/validation/BizValidationCreateTest.kt
@@ -1,33 +1,0 @@
-package space.rybakov.timetable.biz.validation.validation
-
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import space.rybakov.timetable.backend.repository.inmemory.TripRepoStub
-import space.rybakov.timetable.biz.TimetableTripProcessor
-import space.rybakov.timetable.common.TimetableCorSettings
-import space.rybakov.timetable.common.models.TimetableCommand
-import kotlin.test.Test
-
-// TODO-validation-5: смотрим пример теста валидации, собранного из тестовых функций-оберток
-@OptIn(ExperimentalCoroutinesApi::class)
-class BizValidationCreateTest {
-
-    private val command = TimetableCommand.CREATE
-    private val settings by lazy {
-        TimetableCorSettings(
-            repoTest = TripRepoStub()
-        )
-    }
-    private val processor by lazy { TimetableTripProcessor(settings) }
-
-    @Test fun correctName() = validationNameCorrect(command, processor)
-    @Test fun trimName() = validationNameTrim(command, processor)
-    @Test fun emptyName() = validationNameEmpty(command, processor)
-    @Test fun badSymbolsName() = validationNameSymbols(command, processor)
-
-    @Test fun correctDescription() = validationDescriptionCorrect(command, processor)
-    @Test fun trimDescription() = validationDescriptionTrim(command, processor)
-    @Test fun emptyDescription() = validationDescriptionEmpty(command, processor)
-    @Test fun badSymbolsDescription() = validationDescriptionSymbols(command, processor)
-
-}
-
